Moved log gathering to LogReader class
